feat: extract urlBar into component
