+++ src/admin-panel/config.ts
@@ -1,0 +1,9 @@
+export const HTTP_METHODS = [
+	'GET',
+	'HEAD',
+	'POST',
+	'PUT',
+	'DELETE',
+	'OPTIONS',
+	'PATCH',
+]
